+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Removed the default values for `searchFillColor` and `searchCursorColor` (now defaults to `null`).
 * Added style option `listSeparatorColor` to customize the default color of the list separator divider, which has been changed from `Colors.black12` to `Colors.transparent`.
+* Added contextual color models `BrightnessColor` and `ThemedColor` to allow for contextually aware colors.
 
 ## 0.0.1